# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Green Balance - Klima & Wirtschaft API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  /** Total CO2 in million tonnes */
  co2TotalMt: number;
  /** Percentage of renewable energy */
  renewablePercent: number;
  /** Jobs in green industries (thousands) */
  greenJobs: number;
  /** Jobs in traditional energy industries (thousands) */
  traditionalJobs: number;
  /** Electricity price in cents per kWh */
  electricityPriceCents: number;
  /** YoY change in CO2 percentage */
  co2Change: number;
  /** YoY change in renewable percentage */
  renewableChange: number;
  /** YoY change in green jobs percentage */
  greenJobsChange: number;
}

export interface YearlyDataPoint {
  year: number;
  value: number;
  /** @nullable */
  label?: string | null;
}

export interface EnergyDataPoint {
  year: number;
  electricityPrice: number;
  renewableShare: number;
  coalShare: number;
  gasShare: number;
}

export interface JobsDataPoint {
  year: number;
  greenJobs: number;
  coalJobs: number;
  oilGasJobs: number;
}

export type ConflictTopicSeverity = typeof ConflictTopicSeverity[keyof typeof ConflictTopicSeverity];


export const ConflictTopicSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface ConflictTopic {
  id: number;
  title: string;
  /** Positive climate effect description */
  climateEffect: string;
  /** Negative economy effect description */
  economyEffect: string;
  /** Climate benefit 0-10 */
  climateScore: number;
  /** Economy disruption cost 0-10 */
  economyScore: number;
  description: string;
  severity: ConflictTopicSeverity;
}

export interface StateData {
  id: number;
  name: string;
  /** Two-letter state code */
  code: string;
  /** CO2 tonnes per capita */
  co2PerCapita: number;
  renewablePercent: number;
  unemploymentRate: number;
  /** Cents per kWh */
  electricityPrice: number;
  /** Industry as percent of GDP */
  industryShare: number;
  /** Population in thousands */
  population: number;
  /** Installed wind capacity in MW */
  windCapacityMw: number;
}

export interface PollOption {
  id: number;
  text: string;
  votes: number;
}

export interface Poll {
  id: number;
  question: string;
  options: PollOption[];
  totalVotes: number;
}

export interface VoteInput {
  optionId: number;
}

export type DebateRequestRole = typeof DebateRequestRole[keyof typeof DebateRequestRole];


export const DebateRequestRole = {
  unternehmer: 'unternehmer',
  arbeitnehmer: 'arbeitnehmer',
  umweltaktivist: 'umweltaktivist',
  politiker: 'politiker',
} as const;

export interface DebateRequest {
  role: DebateRequestRole;
  /** The climate/economy topic to debate */
  topic: string;
}

export interface DebateResponse {
  role: string;
  /** Full perspective text */
  perspective: string;
  keyPoints: string[];
}

export interface SimulatorInput {
  /**
     * @minimum 0
     * @maximum 100
     */
  windPercent: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  solarPercent: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  coalPercent: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  nuclearPercent: number;
  /**
     * @minimum 0
     * @maximum 500
     */
  co2TaxEuros: number;
}

export interface SimulatorResult {
  co2ReductionPercent: number;
  electricityPriceCents: number;
  jobsCreated: number;
  jobsLost: number;
  renewablePercent: number;
  /** 0-100 score of how feasible/balanced this energy mix is */
  feasibilityScore: number;
  warnings?: string[];
}