# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Green Balance - Klima & Wirtschaft API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /polls/{id}:
    get:
      operationId: getPoll
      tags: [polls]
      summary: Get a specific poll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Poll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "404":
          description: Not found
  /polls/{id}/vote:
    post:
      operationId: votePoll
      tags: [polls]
      summary: Vote on a poll option
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteInput"
      responses:
        "200":
          description: Updated poll after vote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "400":
          description: Invalid option
        "404":
          description: Not found
...
components:
  schemas:
...
    Poll:
      type: object
      required: [id, question, options, totalVotes]
      properties:
        id:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/PollOption"
        totalVotes:
          type: integer
    PollOption:
      type: object
      required: [id, text, votes]
      properties:
        id:
          type: integer
        text:
          type: string
        votes:
          type: integer
    VoteInput:
      type: object
      required: [optionId]
      properties:
        optionId:
          type: integer
...